Netty方式完成

+++ rpc-core/src/main/java/com/tinyrpc/remoting/constants/RpcConstants.java
@@ -25,7 +25,7 @@ public class RpcConstants {
     public static final byte HEARTBEAT_REQUEST_TYPE = 3;
     //pong
     public static final byte HEARTBEAT_RESPONSE_TYPE = 4;
-    public static final byte HEAD_LENGTH = 16;
+    public static final int HEAD_LENGTH = 16;
     public static final String PING = "ping";
     public static final String PONG = "pong";
     public static final int MAX_FRAME_LENGTH = 8 * 1024 * 1024;
